% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{calculate_los}
\alias{calculate_los}
\title{Calculates length of stay for multiple individuals.}
\usage{
calculate_los(occupancy, start_states, times, state_names, ci, start_time = 0)
}
\arguments{
\item{occupancy}{State occupancy data.table as returned by \code{state_occupancy}.}

\item{times}{Times at which to estimate transition probabilities.}

\item{state_names}{Character vector containing the names of the states.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}

\item{end_times}{Times at which to estimate transition probabilities. If not provided then doesn't estimate
transition probabilities, just length of stay.}
}
\value{
A data frame in long format with transition probabilities for each individual,
for each starting time, and for each ending time.
TODO Could make this different to totls.simfs by allowing user to specify starting time as well?
I.e. this currently just assumes state entry at time 0
}
\description{
Uses an already formatted set of arguments to run the individual level
simulation for each new individual (whereas \code{individual_simulation} only accepts 1 person)
and derives transition probabilities from the resulting state occupancies.
}
