% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_probabilities.R
\name{calculate_transition_probabilities}
\alias{calculate_transition_probabilities}
\title{Calculates transition probabilities for multiple individuals.}
\usage{
calculate_transition_probabilities(occupancy, start_times, end_times,
  state_names, ci)
}
\arguments{
\item{times}{Times at which to estimate transition probabilities. If not provided then doesn't estimate
transition probabilities, just length of stay.}

\item{covar_values}{Data frame comprising the covariate values so that neat labels can be formed
for the output probability data frame..}
}
\value{
A data frame in long format with transition probabilities for each individual,
for each starting time, and for each ending time.
}
\description{
Uses an already formatted set of arguments to run the individual level
simulation for each new individual (whereas \code{individual_simulation} only accepts 1 person)
and derives transition probabilities from the resulting state occupancies.
}
