% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coefs_as_list}
\alias{coefs_as_list}
\title{Utility function to obtain covariates as a list for a given distribution, parameterised
by these arguments.}
\usage{
coefs_as_list(all_coefs, param_names, mx, attr_names, dist)
}
\arguments{
\item{all_coefs}{A vector of all the coefficients from a model, including distribution coefficients,
which are located at the start of the vector.}

\item{param_names}{Distribution parameter names as a character vector.}

\item{mx}{The \code{mx} attribute of a \code{flexsurvreg} object, containing indices
of covariates in the coefficient list that act on each distribution parameter.}

\item{attr_names}{Attribute names of the entire attribute matrix used for this model, as a character.
The model may only use a subset of these attributes in each its function, but it needs to know the
indices of the covariates it does use in the wider attribute matrix.}

\item{dist}{The distribution name as a string as used by \code{flexsurvreg}.}
}
\value{
The transition list, as described by \code{obtain_model_coef}.
}
\description{
This function should never be directly called, it is a helper function to reduce code repetition
when making bootstrapped models.
}
