% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort_simulation}
\alias{cohort_simulation}
\title{Runs a cohort discrete event simulation}
\usage{
cohort_simulation(models, newdata, trans_mat, start_times = NULL)
}
\arguments{
\item{models}{Models in the form of a list of \code{flexsurvreg} objects.}

\item{newdata}{Data frame with individual attributes}

\item{trans_mat}{Transition matrix}

\item{start_times}{Entry times of individuals specified in \code{newdata}.
Can either be a single time that everyone enters at, or have as many values
as rows in \code{newdata}. If NULL then everyone enter at t=0}
}
\value{
A data frame with state entry times for each individual.
}
\description{
Runs a cohort discrete event simulation
}
