% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_model_matrix}
\alias{form_model_matrix}
\title{Creates data matrix from data frame and models.}
\usage{
form_model_matrix(dataframe, models)
}
\arguments{
\item{dataframe}{The data to be converted in a standard
\code{data.frame} object.}

\item{models}{List of \code{flexsurvreg} objects.}
}
\value{
A \code{data.matrix} with the data.
}
\description{
Creates data matrix from data frame and models.
}
