% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{individual_simulation}
\alias{individual_simulation}
\title{Runs an individual level simulation.}
\usage{
individual_simulation(transitions, newdata_mat, trans_mat, N, tcovs)
}
\arguments{
\item{transitions}{A list of transitions as required by \code{desCpp}.}

\item{newdata_mat}{New data already in data matrix format as a single row}

\item{trans_mat}{Transition matrix}

\item{N}{Number of times to replicate this individual.}

\item{tcovs}{Indices of covariates that are time-dependent and need to have the time
since simulation start added onto them at each intermediate state entry. Default is NULL.}
}
\value{
A data frame with entry times for each state that is entered.
}
\description{
Returns a data.table with state occupancy times for a single individual.
}
