% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{length_of_stay}
\alias{length_of_stay}
\title{Estimates length of stay}
\usage{
length_of_stay(models, newdata, trans_mat, times, start = 1, tcovs = NULL,
  N = 1e+05, M = 1000, ci = FALSE, ci_margin = 0.95)
}
\arguments{
\item{models}{List of \code{flexsurvreg} objects.}

\item{newdata}{Data frame with covariates of individual to simulate times for. Must contain all fields
required by models.}

\item{trans_mat}{Transition matrix, such as that used in \code{mstate}.}

\item{times}{Times at which to estimate transition probabilities. If not provided then doesn't estimate
transition probabilities, just length of stay.}

\item{start}{Starting state. Either number or character name in \code{trans_mat}.}

\item{tcovs}{As in \code{flexsurv::pmatrix.simfs}, this is the names of covariates that need to be
incremented by the simulation clock at each transition, such as age when modelled as age at state entry.}

\item{N}{Number of times to repeat the individual}

\item{M}{Number of times to run the simulations in order to obtain confidence interval estimates.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}

\item{ci_margin}{Confidence interval range to use if \code{ci} is set to \code{TRUE}.}
}
\value{
A list for each individual with items for length of stay and transition probabilities.
}
\description{
Estimates length of stay in each state of an individual's passage 
through a multi-state model
by discrete event simulation.
}
