% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{multiple_simulations}
\alias{multiple_simulations}
\title{Runs simulations for multiple individuals.}
\usage{
multiple_simulations(transitions, trans_mat, newdata_mat, N, tcovs,
  covar_values)
}
\arguments{
\item{transitions}{A list of transitions as required by \code{desCpp}.}

\item{trans_mat}{Transition matrix}

\item{newdata_mat}{New data already in data matrix format as a single row}

\item{N}{Number of times to replicate this individual.}

\item{tcovs}{Indices of covariates that are time-dependent and need to have the time
since simulation start added onto them at each intermediate state entry. Default is NULL.}

\item{covar_values}{Data frame comprising the covariate values so that neat labels can be formed
for the output probability data frame.}
}
\value{
A data.table with state occupancies and key column 'individual'.
}
\description{
Combines the data.tables for multiple individuals / simulations into
a single table, with appropriate indexing.
}
