% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistateutils.R
\docType{package}
\name{multistateutils}
\alias{multistateutils}
\alias{multistateutils-package}
\title{\code{multistateutils} package.}
\description{
Provides utility functions for already fitted parametric multi-state models,
such as estimating transition probabilities and length of stay, as well as
means of visualising these predictions.
}
