% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_model_coef}
\alias{obtain_model_coef}
\title{Function to obtain the transition list per transition model as required.}
\usage{
obtain_model_coef(mod, attrs, M = 1)
}
\arguments{
\item{mod}{A parametric survival model saved as a \code{flexsurvreg} object.}

\item{attrs}{Attributes of a new individual as a data matrix.}

\item{M}{Number of bootstrap simulations, set to 1 if no bootstrapping is used.}
}
\value{
A list with the following elements:
\itemize{
    \item name The distribution name as a string, as used by \code{flexsurv}.
    \item coefs A nested list, with each child representing a distribution parameter and
      being a list itself, containing 2 items. The first item is the (zero-based) indices in
      \code{attrs} of the model coefficients, and the second contains the coefficients themselves.
}
}
\description{
Function to obtain the transition list per transition model as required.
}
