% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdes.R
\docType{package}
\name{rdes}
\alias{rdes}
\alias{rdes-package}
\title{\code{rdes} package.}
\description{
Provides discrete event simulations for use in estimating transition probabilities from
multi-state models, as well as running whole-cohort simulations.
}
