% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_covariates}
\alias{separate_covariates}
\title{Separates individuals who are indexed by a single column into their constituent covariates.}
\usage{
separate_covariates(dt, covariates)
}
\arguments{
\item{dt}{Input data source. Can be \code{data.table} or \code{data.frame}.}

\item{covariates}{The covariate names that these individuals have, as a character string.
In the description above these are the <covar_name> values.}
}
\value{
A copy of the data with the 'individual' column removed and replaced by a column
  for each covariate in \code{covariates} as a \code{data.frame}.
}
\description{
Requires a data.table or data.frame that has a column called 'individual' that holds identifying
information about that individual combined into a single column for use as a key. The format
of this column is <covar_name>=<covar_value>,<covar2_name>=<covar2_value>...
}
