% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{state_occupancy}
\alias{state_occupancy}
\title{Determines state occupancy for individual level simulations.}
\usage{
state_occupancy(models, trans_mat, newdata, N, tcovs, ci, M)
}
\arguments{
\item{models}{List of \code{flexsurvreg} objects.}

\item{trans_mat}{Transition matrix, such as that used in \code{mstate}.}

\item{newdata}{Data frame with covariates of individual to simulate times for. Must contain all fields
required by models.}

\item{N}{Number of times to repeat the individual}

\item{tcovs}{As in \code{flexsurv::pmatrix.simfs}, this is the names of covariates that need to be
incremented by the simulation clock at each transition, such as age when modelled as age at state entry.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}

\item{M}{Number of times to run the simulations in order to obtain confidence interval estimates.}
}
\value{
A data.table with state occupancies and key columns 'individual' and 'simulation' if appropriate.
}
\description{
Returns a data table with state occupancy times for each individual,
and also simulation if \code{ci} are requested.
}
