% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_probabilities.R
\name{calculate_transition_probabilities}
\alias{calculate_transition_probabilities}
\title{Calculates transition probabilities for multiple individuals.}
\usage{
calculate_transition_probabilities(occupancy, start_times, end_times,
  state_names, ci)
}
\arguments{
\item{occupancy}{State occupancy data.table as returned by \code{state_occupancy}.}

\item{start_times}{Conditional time for transition probability.}

\item{end_times}{Times at which to estimate transition probabilities
transition probabilities, just length of stay.}

\item{state_names}{Character vector containing the names of the states.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}
}
\value{
A data frame in long format with transition probabilities for each individual,
for each starting time, and for each ending time.
}
\description{
Calculates transition probabilities for specified parameters from state
occupancies.
}
