% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{individual_simulation}
\alias{individual_simulation}
\title{Runs an individual level simulation}
\usage{
individual_simulation(models, newdata, trans_mat, N = NULL)
}
\arguments{
\item{N}{Number of times to replicate this individual.}
}
\description{
Runs an individual level simulation
}
