% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{multiple_simulations}
\alias{multiple_simulations}
\title{Runs simulations for multiple individuals.}
\usage{
multiple_simulations(transition_list, trans_mat, newdata_mat, N, tcovs,
  covar_values)
}
\arguments{
\item{trans_mat}{Transition matrix, such as that used in \code{mstate}.}

\item{N}{Number of times to repeat the individual}

\item{tcovs}{As in \code{flexsurv::pmatrix.simfs}, this is the names of covariates that need to be
incremented by the simulation clock at each transition, such as age when modelled as age at state entry.}

\item{covar_values}{Data frame comprising the covariate values so that neat labels can be formed
for the output probability data frame.}
}
\value{
A data.table with state occupancies and key column 'individual'.
}
\description{
Combines the data.tables for multiple individuals / simulations into
a single table, with appropriate indexing.
}
