% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{predict_transitions}
\alias{predict_transitions}
\title{Estimates transition probabilities and length of stay}
\usage{
predict_transitions(models, newdata, trans_mat, times, start_times = 0,
  tcovs = NULL, N = 1e+05, M = 1000, ci = FALSE)
}
\arguments{
\item{models}{List of \code{flexsurvreg} objects.}

\item{newdata}{Data frame with covariates of individual to simulate times for. Must contain all fields
required by models.}

\item{trans_mat}{Transition matrix, such as that used in \code{mstate}.}

\item{times}{Times at which to estimate transition probabilities. If not provided then doesn't estimate
transition probabilities, just length of stay.}

\item{start_times}{Conditional time for transition probability.}

\item{tcovs}{As in \code{flexsurv::pmatrix.simfs}, this is the names of covariates that need to be
incremented by the simulation clock at each transition, such as age when modelled as age at state entry.}

\item{N}{Number of times to repeat the individual}

\item{M}{Number of times to run the simulations in order to obtain confidence interval estimates.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}
}
\value{
A list for each individual with items for length of stay and transition probabilities.
}
\description{
Estimates various measures of an individual's passage through a multi-state model
by discrete event simulation.
}
