% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_probabilities.R
\name{predict_transitions}
\alias{predict_transitions}
\title{Estimates transition probabilities}
\usage{
predict_transitions(models, newdata, trans_mat, times, start_times = 0,
  tcovs = NULL, N = 1e+05, M = 1000, ci = FALSE, ci_margin = 0.95)
}
\arguments{
\item{models}{List of \code{flexsurvreg} objects.}

\item{newdata}{Data frame with covariates of individual to simulate times for. Must contain all fields
required by models.}

\item{trans_mat}{Transition matrix, such as that used in \code{mstate}.}

\item{times}{Times at which to estimate transition probabilities.}

\item{start_times}{Conditional time for transition probability.}

\item{tcovs}{As in \code{flexsurv::pmatrix.simfs}, this is the names of covariates that need to be
incremented by the simulation clock at each transition, such as age when modelled as age at state entry.}

\item{N}{Number of times to repeat the individual}

\item{M}{Number of times to run the simulations in order to obtain confidence interval estimates.}

\item{ci}{Whether to calculate confidence intervals. See \code{flexsurv::pmatrix.simfs} for details.}

\item{ci_margin}{Confidence interval range to use if \code{ci} is set to \code{TRUE}.}
}
\value{
A data frame with estimates of transition probabilities.
}
\description{
Estimates transition probabilities of an individual's passage 
through a multi-state model
by discrete event simulation.
}
\examples{

library(multistateutils)
library(mstate)
library(flexsurv)

# Convert data to long
data(ebmt3)
tmat <- trans.illdeath()
long <- msprep(time=c(NA, 'prtime', 'rfstime'), 
               status=c(NA, 'prstat', 'rfsstat'), 
               data=ebmt3, 
               trans=tmat, 
               keep=c('age', 'dissub'))
               
# Fit parametric models
models <- lapply(1:3, function(i) {
    flexsurvreg(Surv(time, status) ~ age + dissub, data=long, dist='weibull')
})

# New individual to estimate transition probabilities for
newdata <- data.frame(age="20-40", dissub="AML")

# Estimate transition probabilties at 1 year
predict_transitions(models, newdata, tmat, times=365)

# Estimate transition probabilties at 1 year given we know they're alive after 6 months
predict_transitions(models, newdata, tmat, times=365, start_times = 365/2)

}
