% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statepathways.R
\name{state_pathway_flow}
\alias{state_pathway_flow}
\title{Displays the predicted flow for a given individual through the system}
\usage{
state_pathway_flow(models, trans_mat, newdata, times)
}
\arguments{
\item{models}{A list of \code{flexsurv} survival models for
each transition in \code{trans_mat}.}

\item{trans_mat}{A transition matrix.}

\item{newdata}{A data frame containing the attributes of the person
to display the predicted state flow for. As the diagram can only 
be displayed for a single individual it will ignore any rows
after the first.}

\item{times}{The time-points at which to estimate transition
probabilities.}
}
\value{
The HTML widget.
}
\description{
Produces an interactive HTML widget that displays a Sankey diagram showing a 
predicted patient pathway through the multi-state model.
}
